+++ backend/candidates/apps.py
@@ -1,9 +1,0 @@
-from django.apps import AppConfig
-
-
-class CandidatesConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'candidates'
-
-    def ready(self):
-        from . import signals  # noqa: F401